requirements.txt
aiosignal==1.3.1
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
filelock==3.8.2
frozenlist==1.3.3
grpcio==1.51.1
idna==3.4
importlib-resources==5.10.1
jsonschema==4.17.3
MouseInfo==0.1.3
msgpack==1.0.4
numpy==1.24.0
pandas==1.5.2
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==2.6.0
protobuf==4.21.12
PyAudio==0.2.12
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.19.2
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2022.7
PyYAML==6.0
ray==2.2.0
requests==2.28.1
screen-recorder-sdk==1.3.0
six==1.16.0
urllib3==1.26.13
virtualenv==20.17.1
Wave==0.0.2
zipp==3.11.0
